Move various components to tool_panel folder
