+++ app/frontend/components/tool_panel/commentary_panel.js
@@ -16,8 +16,8 @@
    along with Ezra Bible App. See the file LICENSE.
    If not, see <http://www.gnu.org/licenses/>. */
 
-const eventController = require('../controllers/event_controller.js');
-const VerseBox = require("../ui_models/verse_box.js");
+const eventController = require('../../controllers/event_controller.js');
+const VerseBox = require("../../ui_models/verse_box.js");
 
 /**
  * The CommentaryPanel component implements a tool panel that shows Bible commentaries for selected verses
