+++ app/frontend/components/tool_panel/dictionary_info_box.js
@@ -16,8 +16,8 @@
    along with Ezra Bible App. See the file LICENSE.
    If not, see <http://www.gnu.org/licenses/>. */
 
-const eventController = require('../controllers/event_controller.js');
-const { html } = require('../helpers/ezra_helper.js');
+const eventController = require('../../controllers/event_controller.js');
+const { html } = require('../../helpers/ezra_helper.js');
 
 let jsStrongs = null;
