+++ app/frontend/components/tool_panel/tag_statistics.js
