+++ app/frontend/components/tool_panel/translation_comparison.js
@@ -16,9 +16,9 @@
    along with Ezra Bible App. See the file LICENSE.
    If not, see <http://www.gnu.org/licenses/>. */
 
-const VerseBox = require("../ui_models/verse_box.js");
-const i18nHelper = require('../helpers/i18n_helper.js');
-const eventController = require('../controllers/event_controller.js');
+const VerseBox = require("../../ui_models/verse_box.js");
+const i18nHelper = require('../../helpers/i18n_helper.js');
+const eventController = require('../../controllers/event_controller.js');
 
 /**
  * The TranslationComparison component implements a tool panel that shows selected verses
