+++ app/frontend/controllers/app_controller.js
@@ -23,7 +23,7 @@ const VerseBoxHelper = require("../helpers/verse_box_helper.js");
 const VerseSelection = require("../components/verse_selection.js");
 const VerseListPopup = require("../components/verse_list_popup.js");
 const TagSelectionMenu = require("../components/tags/tag_selection_menu.js");
-const TagStatistics = require("../components/tags/tag_statistics.js");
+const TagStatistics = require("../components/tool_panel/tag_statistics.js");
 const DocxExport = require("../components/docx_export/docx_export.js");
 const ModuleSearchController = require("./module_search_controller.js");
 const TranslationController = require("./translation_controller.js");
@@ -33,8 +33,8 @@ const TabSearch = require("../components/tab_search/tab_search.js");
 const TabController = require("./tab_controller.js");
 const OptionsMenu = require("../components/options_menu/options_menu.js");
 const NavigationPane = require("../components/navigation_pane.js");
-const TranslationComparison = require("../components/translation_comparison.js");
-const CommentaryPanel = require("../components/commentary_panel.js");
+const TranslationComparison = require("../components/tool_panel/translation_comparison.js");
+const CommentaryPanel = require("../components/tool_panel/commentary_panel.js");
 const BookSelectionMenu = require("../components/book_selection_menu.js");
 const DictionaryController = require("./dictionary_controller.js");
 const NotesController = require("./notes_controller.js");
